Extract grid item into its own component & fragment container, for impending reuse.
