+++ src/Components/Search/Previews/Grids/MerchandisableArtworks.tsx
@@ -1,12 +1,13 @@
 import React from "react"
 
-import { Box, Flex, Image, Link, Sans, Serif } from "@artsy/palette"
+import { Box, Flex, Sans } from "@artsy/palette"
 import { MerchandisableArtworks_viewer } from "__generated__/MerchandisableArtworks_viewer.graphql"
 import { MerchandisableArtworksPreviewQuery } from "__generated__/MerchandisableArtworksPreviewQuery.graphql"
 import { renderWithLoadProgress } from "Artsy/Relay/renderWithLoadProgress"
 import { ContextConsumer, ContextProps } from "Artsy/SystemContext"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 import { get } from "Utils/get"
+import { PreviewGridItemFragmentContainer as PreviewGridItem } from "./PreviewGridItem"
 
 interface MerchandisableArtworksPreviewProps {
   viewer: MerchandisableArtworks_viewer
@@ -22,7 +23,7 @@ const MerchandisableArtworksPreview: React.SFC<
   ).map(x => x.node)
 
   const merchandisableItems = artworks.map((artwork, i) => (
-    <MerchandisableArtworkItem artwork={artwork} key={i} />
+    <PreviewGridItem artwork={artwork} key={i} />
   ))
 
   return (
@@ -36,25 +37,6 @@ const MerchandisableArtworksPreview: React.SFC<
   )
 }
 
-const MerchandisableArtworkItem = ({ artwork }) => {
-  const imageUrl = get(artwork, x => x.image.cropped.url, "")
-  return (
-    <Flex m={2}>
-      <Link href={artwork.href} noUnderline>
-        <Image mr={2} src={imageUrl} />
-      </Link>
-      <Link href={artwork.href} noUnderline>
-        <Box>
-          <Serif size="2" italic>
-            {artwork.title}, {artwork.date}
-          </Serif>
-          <Serif size="2">{artwork.artist_names}</Serif>
-        </Box>
-      </Link>
-    </Flex>
-  )
-}
-
 export const MerchandisableArtworksPreviewFragmentContainer = createFragmentContainer(
   MerchandisableArtworksPreview,
   graphql`
@@ -64,15 +46,7 @@ export const MerchandisableArtworksPreviewFragmentContainer = createFragmentCont
         artworks_connection(first: 8) {
           edges {
             node {
-              href
-              title
-              artist_names
-              image {
-                cropped(width: 40, height: 40) {
-                  url
-                }
-              }
-              date
+              ...PreviewGridItem_artwork
             }
           }
         }
